* Refactor common annotations and reuse in molecule entity
